Extracting the whole serializing functionality to a concern
We intend to reuse it (as much as we can) in the Translation class, but
first we are checking that no test breaks in the process
